+++ src/ArchiveFormats/XP3/Crypt/Crypt.h
@@ -8,9 +8,9 @@ class XP3Crypt {
         bool StartupTjsNotEncrypted = false;
         bool ObfuscatedIndex = false;
 
-        virtual std::vector<uint8_t> Decrypt(Entry *entry, uint32_t offset, std::vector<uint8_t> buffer, int pos, int count) = 0;
-        virtual uint8_t Decrypt(Entry *entry, uint32_t offset, uint8_t value) = 0;
-        virtual uint8_t Encrypt(Entry *entry, uint32_t offset, uint8_t value) = 0;
+        virtual std::vector<uint8_t> Decrypt(Entry *entry, uint64_t offset, std::vector<uint8_t> buffer, int pos, int count) = 0;
+        virtual uint8_t Decrypt(Entry *entry, uint64_t offset, uint8_t value) = 0;
+        virtual uint8_t Encrypt(Entry *entry, uint64_t offset, uint8_t value) = 0;
 
         virtual ~XP3Crypt() = default;
 
@@ -46,15 +46,15 @@ class XP3Crypt {
 
 class NoCrypt : public XP3Crypt {
     public:
-        std::vector<uint8_t> Decrypt(Entry *entry, uint32_t offset, std::vector<uint8_t> buffer, int pos, int count) override {
+        std::vector<uint8_t> Decrypt(Entry *entry, uint64_t offset, std::vector<uint8_t> buffer, int pos, int count) override {
             Logger::log("NoCrypt: Decrypting %d bytes at offset %d", count, offset);
             return buffer;
         }
 
-        uint8_t Decrypt(Entry *entry, uint32_t offset, uint8_t value) override {
+        uint8_t Decrypt(Entry *entry, uint64_t offset, uint8_t value) override {
             return value;
         }
-        uint8_t Encrypt(Entry *entry, uint32_t offset, uint8_t value) override {
+        uint8_t Encrypt(Entry *entry, uint64_t offset, uint8_t value) override {
             return value;
         }
 
@@ -65,14 +65,14 @@ class NoCrypt : public XP3Crypt {
 
 class HibikiCrypt : XP3Crypt {
     public:
-        uint8_t Decrypt(Entry *entry, uint32_t offset, uint8_t value) override {
+        uint8_t Decrypt(Entry *entry, uint64_t offset, uint8_t value) override {
             if (0 != (offset & 4) || offset <= 0x64)
                 return (uint8_t)(value ^ (entry->hash >> 5));
             else
                 return (uint8_t)(value ^ (entry->hash >> 8));
         }
 
-        std::vector<uint8_t> Decrypt(Entry *entry, uint32_t offset, std::vector<uint8_t> buffer, int pos, int count) override {
+        std::vector<uint8_t> Decrypt(Entry *entry, uint64_t offset, std::vector<uint8_t> buffer, int pos, int count) override {
             uint8_t key1 = (uint8_t)(entry->hash >> 5);
             uint8_t key2 = (uint8_t)(entry->hash >> 8);
             for (int i = 0; i < count; i++) {
@@ -86,7 +86,7 @@ class HibikiCrypt : XP3Crypt {
         }
 
         // no-op
-        uint8_t Encrypt(Entry *entry, uint32_t offset, uint8_t value) override {
+        uint8_t Encrypt(Entry *entry, uint64_t offset, uint8_t value) override {
             return value;
         }
